app/requirements.txt
Django==4.1.4
django-admin-sortable2==2.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-jazzmin==2.6.0
django-quill-editor==0.1.40
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dj_database_url
drf-nested-routers==0.93.4
drf-yasg==1.21.4
gunicorn==20.1.0
hashids==1.3.1
pillow==9.3.0
psycopg2-binary==2.9.5
pylint==2.15.8
pylint-django==2.5.3
shortuuid==1.0.11
